Factor classes out into separate files
