+++ pages/_includes/mlayout.hbs
@@ -10,7 +10,7 @@
     <link rel="dns-prefetch" href="https://fonts.googleapis.com/">
     <meta name="description" content="Gain practical experience by working remotely on real world projects with other early-career developers.">
     <link href="/assets/styles.css" rel="stylesheet">
-    <script defer src="/assets/behaviors.js"></script>
+    <script defer type="module" src="/assets/behaviors.js"></script>
     <script type="application/ld+json">
     {
       "@context": "http://www.schema.org",
